+++ src/main/java/az/m10/dto/TotalMealValuesDTO.java
@@ -2,17 +2,15 @@ package az.m10.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class TotalMealValuesDTO {
-    private BigDecimal totalProtein;
-    private BigDecimal totalCarbs;
-    private BigDecimal totalCal;
-    private BigDecimal totalFat;
-    private BigDecimal totalSugar;
+    private Double totalProtein;
+    private Double totalCarbs;
+    private Double totalCal;
+    private Double totalFat;
+    private Double totalSugar;
 }
